nemo_langgraph_project_working/requirements.txt
# Minimal requirements for AIQ + LangGraph + NVIDIA NeMo projects
# This file contains only the essential packages for core functionality

# Core AIQ packages
aiqtoolkit>=1.1.0
aiqtoolkit-langchain>=1.1.0

# Environment and configuration
python-dotenv>=1.0.0
pydantic>=2.0.0,<3.0
PyYAML>=6.0

# LangChain ecosystem  
langchain>=0.3.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
langchain-nvidia-ai-endpoints>=0.3.0
langchain-text-splitters>=0.3.0

# LangGraph
langgraph>=0.2.0
langgraph-checkpoint>=2.0.0

# LangSmith for tracing
langsmith>=0.4.0

# HTTP and API clients
httpx>=0.24.0
requests>=2.25.0,<3.0
aiohttp>=3.8.0

# Data handling
pandas>=2.0.0
numpy>=1.24.0
datasets>=4.0.0

# ML/AI libraries
transformers>=4.30.0
torch>=2.0.0
sentence-transformers>=3.0.0

# Jupyter for development
jupyter>=1.0.0
ipykernel>=6.0.0

# Development tools
pytest>=7.0.0
black>=23.0.0
python-multipart>=0.0.6

# FastAPI for web interfaces
fastapi>=0.100.0
uvicorn>=0.20.0

# CLI and utilities
click>=8.0.0
rich>=13.0.0
typer>=0.9.0
tqdm>=4.65.0

# Async support
asyncio-throttle>=1.0.0
nest-asyncio>=1.5.0

# File and data formats
openpyxl>=3.1.0
pypdf>=5.0.0
beautifulsoup4>=4.12.0

# Monitoring and telemetry (optional but useful)
mlflow>=2.0.0
wandb>=0.16.0

# OpenTelemetry for metrics and tracing
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-instrumentation-langchain